refactor: node/web/common file structure (#5336)
* doc: update shared, common, web, node subfolder structure


* refactor: move `crypto` module to `shared/`

We are moving everything out of the `common/` folder
since we do not need this anymore.


* refactor: move `request` in to `shared/`

This removes all modules from the `common/` folder.


* doc: update web docs


* refactor: export a Node only codewhisperer module

Problem:

We have some code, specifically Transform/Gumby, that only works
in Node.js. Then in `codewhisperer/index.ts` we export many CW related
code including this gumby code.

The issue arises when we import this `index.ts` in to Web only code
as it transitively imports the gumby code and this breaks due to
them using fs-extra.

Solution:

Create a `codewhisperer/indexNode.ts` file and a `codewhisperer/index.ts` file.
- `indexNode` is a superset of `index`.
- `indexNode` will additionally include Node only code
- `index` will include "common" code, meaning it works in Web or Node
  - Note this is not 100% true yet as there may still be node only code in this. We will just need to move it to `indexNode` over time


* refactor: remove redundant exports

There were some redundant exports that used "as".
This removes them and updates to use the direct import
instead.


* refactor: extension.ts -> extensionNode.ts

Adds the "node" keyword to the extension file
and updates exports/imports.

We want "extension.ts" to represent common code going
forward.


* refactor: extensionCommon.ts -> extension.ts

Now the extension.ts files will implicitly be "common" files
and extensionNode will contain Node specific code


* minor rename


* minor doc fix


---------
