+++ packages/amazonq/package.json
@@ -42,7 +42,7 @@
         "onCommand:aws.amazonq.accept",
         "onView:aws.codeWhisperer.securityPanel"
     ],
-    "main": "./dist/src/extension",
+    "main": "./dist/src/extensionNode",
     "browser": "./dist/src/extensionWeb",
     "scripts": {
         "vscode:prepublish": "npm run clean && npm run buildScripts && webpack --mode production",
