+++ packages/amazonq/test/unit/amazonqGumby/resources/mocks/transformFixtures.ts
@@ -2,7 +2,7 @@
  * Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
  * SPDX-License-Identifier: Apache-2.0
  */
-import { TransformationStep } from 'aws-core-vscode/codewhisperer'
+import { TransformationStep } from 'aws-core-vscode/codewhisperer/node'
 
 export const downloadArtifactIdFixture = 'hil-test-artifact-id'
 export const downloadArtifactTypeFixture = 'BuiltJars'
