+++ packages/amazonq/webpack.config.js
@@ -10,7 +10,7 @@ module.exports = (env, argv) => {
     const config = {
         ...baseConfigFactory(env, argv),
         entry: {
-            'src/extension': './src/extension.ts',
+            'src/extensionNode': './src/extensionNode.ts',
         },
     }
