+++ packages/core/src/codewhisperer/index.ts
@@ -3,11 +3,10 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-export * from './indexCommon'
-
+export { activate, shutdown } from './activation'
+export * from './util/authUtil'
 export * from './models/model'
-export * as model from './models/model'
-export * as CodeWhispererConstants from './models/constants'
+export * from './models/constants'
 export * from './commands/basicCommands'
 export * from './commands/types'
 export {
@@ -28,17 +27,8 @@ export type {
     TelemetryEvent,
 } from './client/codewhispereruserclient.d.ts'
 export type { default as CodeWhispererUserClient } from './client/codewhispereruserclient.d.ts'
-export { HumanInTheLoopManager } from './service/transformByQ/humanInTheLoopManager'
-export * from './service/transformByQ/transformApiHandler'
-export {
-    DiffModel,
-    AddedChangeNode,
-    ModifiedChangeNode,
-} from './service/transformByQ/transformationResultsViewProvider'
-export { parseVersionsListFromPomFile } from './service/transformByQ/transformFileHandler'
 export { SecurityPanelViewProvider } from './views/securityPanelViewProvider'
 export { isInlineCompletionEnabled } from './util/commonUtil'
-export { validateOpenProjects, getOpenProjects } from './service/transformByQ/transformProjectValidationHandler'
 export {
     DefaultCodeWhispererClient,
     Recommendation,
