+++ packages/core/src/codewhisperer/indexNode.ts
@@ -1,0 +1,22 @@
+/*!
+ * Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+/**
+ * This module is a superset of index.ts. Index.ts is for "common" code.
+ *
+ * The only exports that should be set in this module are those that only run
+ * in Node.js environments.
+ */
+export * from './index'
+
+export { HumanInTheLoopManager } from './service/transformByQ/humanInTheLoopManager'
+export * from './service/transformByQ/transformApiHandler'
+export {
+    DiffModel,
+    AddedChangeNode,
+    ModifiedChangeNode,
+} from './service/transformByQ/transformationResultsViewProvider'
+export { parseVersionsListFromPomFile } from './service/transformByQ/transformFileHandler'
+export { validateOpenProjects, getOpenProjects } from './service/transformByQ/transformProjectValidationHandler'
