+++ packages/core/src/test/index.ts
@@ -24,7 +24,7 @@ export * from './credentials/testUtil'
 export * from './testUtil'
 export * from './amazonqFeatureDev/utils'
 
-import request from '../common/request'
+import request from '../shared/request'
 import { stub } from 'sinon'
 
 // Returns a stubbed fetch for other tests.
